Fix NPE when checking for percentage changes
Fix NPE when checking for percentage changes

     Fix NPE when checking for percentage changes with small map sizes.

     Fix error on player-join when join-message is null.
